Load VO mapping in correct place
